Display unescaped strings for the relevant trailer document fields
When trailer information is stored in the db, strings are escaped
for security reasons. When the documents are retrieved, there
needs to be a way of unescaping the strings so they display
nicely. This has been done by implementing virtual getter functions
on the trailer model and leaning on the he npm module for the
decoding itself.