Add modal overlay to trailers to watch, showing ratings window
A modal overlay has been added to the trailers to watch view, that
is displayed when you click the rate film button for any of the
trailers. The title of the modal window is the title of the film
that you are rating.

The db id of the trailer still needs to be passed into the window,
as this will needed to actually update the database with the
chosen rating.