Rework the email trailer import script
The script that imports trailers from Thunderbird was completely
reworked to improve readability, fix bugs and simplify the code.

Note there is still a bug in the code that means you can't
currently close the database connection programmatically without
it crashing. This prevents any database reads from taking place.

The workaround for this at the moment is not to close the
connection at all, instead closing it manually with Ctrl+C once
the script has finished running.