Add 'Teaser' as a filter check on the tags for downloaded trailers
Trailers that are downloaded pass through a filter to remove
anything we're not interested in. The filter was already removing
trailers with the word 'teaser' in the title, however some have
the word in the tags instead. The filter was updated to remove
trailers with 'teaser' as one of their tags.